+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Added original comments wrapper. Should help themes align comments properly, #283
 * Added "WordPress only" as a form type in "Elements" tab
+* Added new option under "Generic" tab to define comment update time in minutes
 
 **Fixes:**